Update Position struct to use usize instead of u64.

+++ src/buffer/position.rs
@@ -1,5 +1,5 @@
 #[derive(Copy, Clone)]
 pub struct Position {
-    pub line:   u64,
-    pub offset: u64,
+    pub line:   usize,
+    pub offset: usize,
 }
